First draft of DI container object
The intention is that this object will be passed to the
constructor of each node, delegating responsibility for
selecting the correct channel tx/rx ends and singletons
to each node's implementation.

The current `new()` function for DiContainer accepts
a pre-constructed HashMap with all the required channels.
I will need to identify the best place to convert a set
of node definitions into this set of channels, to avoid
tight coupling.